+++ src/webgis4u/ol/Settings.js
@@ -1,0 +1,12 @@
+
+const BOUNDINGBOX_BUFFER = 2500;
+
+const Settings = {
+  /**
+   * The default buffer that is added to an extent with the width and height = 0 (The default value is 2500m).
+   */
+  BOUNDINGBOX_BUFFER,
+};
+Object.freeze(Settings);
+
+export default Settings;