+++ src/webgis4u/ol/util.js
@@ -2,5 +2,11 @@
  * @module webgis4u/ol/util
  */
 
+export * from './util/createImageFromMap';
+export * from './util/getMapChildElementByClassName';
 export * from './util/getMapType';
 export * from './util/getScale';
+export * from './util/pan';
+export * from './util/print';
+export * from './util/zoomToExtent';
+export * from './util/zoomToLayerExtent';