+++ src/webgis4u/util/ol.js
@@ -1,7 +1,0 @@
-/**
- * @fileoverview contains various util functions related to OpenLayers
- * @module webgis4u/util/ol
- */
-
-export * from './ol/getMapChildElementByClassName';
-export * from './ol/pan';